feat: added media to the structure of the product caed
