--- src/entities/product-card/ui/media/index.ts
@@ -1,1 +1,0 @@
-export * as Media from "./media";
